+++ cpp/libjoynr/joynr-messaging/AbstractJoynrMessagingConnector.cpp
@@ -30,7 +30,7 @@ AbstractJoynrMessagingConnector::AbstractJoynrMessagingConnector(
         ISubscriptionManager* subscriptionManager,
         const std::string& domain,
         const std::string& interfaceName,
-        const std::string proxyParticipantId,
+        const std::string& proxyParticipantId,
         const std::string& providerParticipantId,
         const MessagingQos& qosSettings,
         IClientCache* cache,